// Revision Signal r/SCI_Client-SignalR (1)/SCI_Client-SignalR/ngsci-depart-main/src/app/models/models.ts
export interface Player {
    id: number;
    name: string;
}

export interface Card {
    id: number;
    name: string;
    attack: number;
    health: number;
    cost: number;
    imageUrl: string;
}

export interface MatchData {
    match:Match;
    playerA: Player;
    playerB: Player;
    winningPlayerId:number;
}

export interface Match {
    id: number;
    isMatchCompleted: boolean;
    isPlayerATurn: boolean;
    playerDataA: PlayerData;
    playerDataB: PlayerData;
}

export interface PlayableCard {
    id: number;
    card: Card;
    health: number;
}

export interface PlayerData {
    id:number;
	health: number;
    maxhealth: number;
    mana: number;
    playerId: number;
    playerName: string;
    cardsPile: PlayableCard[];
    hand: PlayableCard[];
    battleField: PlayableCard[];
    graveyard: PlayableCard[];
}

//JoinMatchData depuis la fonction JoinMatch SignalR
export interface JoinMatchData{
    match:Match;
    playerA:Player;
    playerB:Player;
    isStarted:boolean;
    otherPlayerConnectionId:string;
}